requirements.txt
python-box
pyYAML
ensure
joblib
types-PyYAML
python-dotenv
pandas
psycopg2
ipykernel